+++ NEWS.md
@@ -3,7 +3,10 @@
 # ApplyPolygenicScore unreleased
 ## Changed
 * Fixed regression of combine.vcf.with.pgs() function that prevented it from handling multiple rsIDs on the same line.
-* added new contributor
+* Fixed bug caused by the case of a sample-specific missing variant at a multiallelic site
+
+## Added
+* Added new contributor
 
 # ApplyPolygenicScore 3.0.2